+++ src/DepthTensor/__init__.py
@@ -6,4 +6,4 @@ from ._core.ops import *
 from ._core.exceptions import *
 from ._core import random
 
-__version__ = "2.2.1"
+__version__ = "2.3.0"